Add Unmanaged classes for all GPU/audio resources
