+++ include/raylib-cpp.hpp
@@ -33,6 +33,7 @@
 
 #include "./AudioDevice.hpp"
 #include "./AudioStream.hpp"
+#include "./AudioStreamUnmanaged.hpp"
 #include "./AutomationEventList.hpp"
 #include "./BoundingBox.hpp"
 #include "./Camera2D.hpp"
@@ -41,17 +42,21 @@
 #include "./FileData.hpp"
 #include "./FileText.hpp"
 #include "./Font.hpp"
+#include "./FontUnmanaged.hpp"
 #include "./Functions.hpp"
 #include "./Gamepad.hpp"
 #include "./Image.hpp"
 #include "./Keyboard.hpp"
 #include "./Material.hpp"
+#include "./MaterialUnmanaged.hpp"
 #include "./Matrix.hpp"
 #include "./Mesh.hpp"
 #include "./Model.hpp"
+#include "./ModelUnmanaged.hpp"
 #include "./ModelAnimation.hpp"
 #include "./Mouse.hpp"
 #include "./Music.hpp"
+#include "./MusicUnmanaged.hpp"
 #include "./Ray.hpp"
 #include "./RayCollision.hpp"
 #include "./RaylibException.hpp"
@@ -60,6 +65,7 @@
 #include "./RenderTextureUnmanaged.hpp"
 #include "./Shader.hpp"
 #include "./Sound.hpp"
+#include "./SoundUnmanaged.hpp"
 #include "./Text.hpp"
 #include "./Texture.hpp"
 #include "./TextureUnmanaged.hpp"
@@ -69,6 +75,7 @@
 #include "./Vector4.hpp"
 #include "./VrStereoConfig.hpp"
 #include "./Wave.hpp"
+#include "./WaveUnmanaged.hpp"
 #include "./Window.hpp"
 
 /**
